sync: Fix loop and stop using uint8_t as size type
The typo in the loop was not discovered because after running for 256
iterations the loop terminated due to uint8_t wrap around. Using
uint32_t makes similar issues more noticeable.

We always had this typo. Not sure which scenarios this could affect,
but as the first step this should be fixed.

Not the first issue with uint8_t we encountered, so start using
uint32_t as default size type for small_vector. This does not affect
the final size of the container due to alignment gaps. For specific
case of small_vector<ResourceFirstAccess, 3> it's 96 bytes both for
uint8_t and uint32_t size type parameterization.

+++ layers/containers/custom_containers.h
@@ -1,6 +1,6 @@
-/* Copyright (c) 2015-2017, 2019-2023 The Khronos Group Inc.
- * Copyright (c) 2015-2017, 2019-2023 Valve Corporation
- * Copyright (c) 2015-2017, 2019-2023 LunarG, Inc.
+/* Copyright (c) 2015-2017, 2019-2024 The Khronos Group Inc.
+ * Copyright (c) 2015-2017, 2019-2024 Valve Corporation
+ * Copyright (c) 2015-2017, 2019-2024 LunarG, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -118,7 +118,7 @@ using insert_iterator = std::insert_iterator<T>;
 //       MoveAssignable and MoveConstructable
 // NOTE: Unlike std::vector, iterators are invalidated by move assignment between small_vector objects effectively the
 //       "small string" allocation functions as an incompatible allocator.
-template <typename T, size_t N, typename SizeType = uint8_t>
+template <typename T, size_t N, typename SizeType = uint32_t>
 class small_vector {
   public:
     using value_type = T;
